transaction_service/requirements.txt
Click==7.0
Flask==1.1.1
grpcio==1.25.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
protobuf==3.10.0
python-dotenv==0.10.3
six==1.13.0
Werkzeug==0.16.0
